test/owning-a-home/api_testing/requirements.txt
selenium==2.40.0
nose
pyhamcrest==1.8.0
behave==1.2.4
